refactor(utils): rename packing to proof-packing
re #177
