+++ packages/utils/src/index.ts
@@ -1,13 +1,13 @@
 import F1Field from "./f1-field"
 import * as scalar from "./scalar"
 import * as conversions from "./conversions"
-import * as packing from "./packing"
+import * as packing from "./proof-packing"
 import * as typeChecks from "./type-checks"
 import * as errorHandlers from "./error-handlers"
 
 export { F1Field, scalar, conversions, packing, typeChecks, errorHandlers }
 export * from "./types"
 export * from "./conversions"
-export * from "./packing"
+export * from "./proof-packing"
 export * from "./type-checks"
 export * from "./error-handlers"
